refactor(tech-tree): split node and legend rendering parts
Reduce tech_tree_widget.dart size by extracting node rendering and legend widgets into part files with no behavior change, completing another maintainability slice for issue #1878.

Made-with: Cursor
