+++ app/lib/features/game/widgets/tech_tree_widget.dart
@@ -15,6 +15,9 @@ import '../../../widgets/ct_nine_patch_button.dart';
 import '../../../widgets/strict_asset_icon.dart';
 import 'tech_effect_summary_lookup.dart';
 
+part 'tech_tree_widget_nodes.dart';
+part 'tech_tree_widget_legend.dart';
+
 /// Node position for layout. Exposed for tests (column rule: A→B→C and A→C ⇒ gap between A and C).
 class TechNodePosition {
   const TechNodePosition({
@@ -330,9 +333,7 @@ class TechTreeWidget extends StatelessWidget {
                 (id) => Padding(
                   padding: const EdgeInsets.only(top: 2),
                   child: Text(
-                    l10n.techTree_prerequisiteBullet(
-                      techDisplayName(id),
-                    ),
+                    l10n.techTree_prerequisiteBullet(techDisplayName(id)),
                     style: theme.textTheme.bodySmall,
                   ),
                 ),
@@ -340,10 +341,7 @@ class TechTreeWidget extends StatelessWidget {
             ],
             if (effects.isNotEmpty) ...[
               const SizedBox(height: 8),
-              Text(
-                l10n.techTree_effects,
-                style: theme.textTheme.labelLarge,
-              ),
+              Text(l10n.techTree_effects, style: theme.textTheme.labelLarge),
               ...effects.map(
                 (e) => Padding(
                   padding: const EdgeInsets.only(top: 2),
@@ -421,322 +419,5 @@ class TechTreeWidget extends StatelessWidget {
               : '${s[0].toUpperCase()}${s.substring(1).toLowerCase()}',
         )
         .join(' ');
-  }
-}
-
-class _TechNodeState {
-  const _TechNodeState({
-    required this.researched,
-    required this.inProgress,
-    required this.available,
-  });
-  final bool researched;
-  final bool inProgress;
-  final bool available;
-}
-
-class _TechTreeEdgePainter extends CustomPainter {
-  _TechTreeEdgePainter({required this.positions});
-
-  final List<TechNodePosition> positions;
-
-  static double get _centerY => _nodeHeight / 2;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final posByTech = {for (final p in positions) p.techId: p};
-    final paint = Paint()
-      ..color = Colors.grey.shade600
-      ..strokeWidth = _edgeStrokeWidth
-      ..style = PaintingStyle.stroke;
-
-    for (final tech in techCatalog.values) {
-      final toPos = posByTech[tech.id];
-      if (toPos == null) continue;
-      final toLeftX = toPos.x;
-      final toCenterY = toPos.y + _centerY;
-      for (final prereqId in tech.prerequisiteIds) {
-        final fromPos = posByTech[prereqId];
-        if (fromPos == null) continue;
-        final fromRightX = fromPos.x + _nodeWidth;
-        final fromCenterY = fromPos.y + _centerY;
-
-        // Right-angled connector: horizontal into gap, vertical to target row, horizontal to target.
-        // Layout reserves a row slot in intermediate columns so this segment does not pass through nodes.
-        final bendX = fromRightX + _edgeBendOffset;
-        final path = Path()
-          ..moveTo(fromRightX, fromCenterY)
-          ..lineTo(bendX, fromCenterY)
-          ..lineTo(bendX, toCenterY)
-          ..lineTo(toLeftX, toCenterY);
-        canvas.drawPath(path, paint);
-      }
-    }
-  }
-
-  @override
-  bool shouldRepaint(covariant CustomPainter oldDelegate) => false;
-}
-
-class _TechNode extends StatelessWidget {
-  const _TechNode({
-    required this.tech,
-    required this.state,
-    required this.onTap,
-  });
-
-  final TechDefinition tech;
-  final _TechNodeState state;
-  final VoidCallback onTap;
-
-  @override
-  Widget build(BuildContext context) {
-    final style = _nodeStyle();
-    return Material(
-      color: Colors.transparent,
-      child: InkWell(
-        onTap: onTap,
-        child: Container(
-          width: _nodeWidth,
-          height: _nodeHeight,
-          decoration: BoxDecoration(
-            color: style.fillColor,
-            border: Border.all(color: style.borderColor, width: style.borderWidth),
-            borderRadius: BorderRadius.circular(6),
-          ),
-          child: Center(
-            child: Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 4),
-              child: _buildNodeLabel(style.locked),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-
-  _TechNodeStyle _nodeStyle() {
-    final color = _categoryColors[tech.category] ?? Colors.grey;
-    final locked = !state.researched && !state.inProgress && !state.available;
-    if (state.researched) {
-      return _TechNodeStyle(
-        fillColor: color,
-        borderColor: color.withValues(alpha: 0.8),
-        borderWidth: 2,
-        locked: locked,
-      );
-    }
-    if (state.inProgress) {
-      return _TechNodeStyle(
-        fillColor: color.withValues(alpha: 0.4),
-        borderColor: color,
-        borderWidth: 3,
-        locked: locked,
-      );
-    }
-    if (state.available) {
-      return _TechNodeStyle(
-        fillColor: color.withValues(alpha: 0.15),
-        borderColor: color,
-        borderWidth: 2,
-        locked: locked,
-      );
-    }
-    return _TechNodeStyle(
-      fillColor: Colors.grey.shade200,
-      borderColor: Colors.grey.shade400,
-      borderWidth: 1,
-      locked: locked,
-    );
-  }
-
-  Widget _buildNodeLabel(bool locked) {
-    return Row(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        if (_categoryIcons.containsKey(tech.category))
-          Padding(
-            padding: const EdgeInsets.only(right: 3),
-            child: StrictAssetIcon(
-              assetPath: _categoryIcons[tech.category]!,
-              width: 16,
-              height: 16,
-            ),
-          ),
-        Flexible(
-          child: Text(
-            techDisplayName(tech.id),
-            textAlign: TextAlign.center,
-            maxLines: 2,
-            overflow: TextOverflow.ellipsis,
-            style: TextStyle(
-              fontSize: 10,
-              color: locked ? Colors.grey : null,
-              fontWeight: state.researched ? FontWeight.w600 : null,
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class _TechNodeStyle {
-  const _TechNodeStyle({
-    required this.fillColor,
-    required this.borderColor,
-    required this.borderWidth,
-    required this.locked,
-  });
-
-  final Color fillColor;
-  final Color borderColor;
-  final double borderWidth;
-  final bool locked;
-}
-
-/// Row label for tech tree legend samples (maps to [AppLocalizations] state strings).
-enum _TechLegendStateKind { researched, inProgress, available, locked }
-
-class _TechTreeLegend extends StatelessWidget {
-  const _TechTreeLegend({required this.l10n});
-
-  final AppLocalizations l10n;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(l10n.techTree_legendTitle, style: theme.textTheme.labelLarge),
-        const SizedBox(height: 4),
-        _buildCategoryLegendWrap(),
-        const SizedBox(height: 8),
-        _buildStateLegendWrap(),
-      ],
-    );
-  }
-
-  Widget _buildCategoryLegendWrap() {
-    return Wrap(
-      spacing: 8,
-      runSpacing: 4,
-      children: _categoryColors.entries
-          .map(
-            (e) => _LegendChip(
-              color: e.value,
-              label: TechTreeWidget._categoryLabelL10n(l10n, e.key),
-            ),
-          )
-          .toList(),
-    );
-  }
-
-  Widget _buildStateLegendWrap() {
-    return Wrap(spacing: 8, runSpacing: 4, children: _stateLegendSamples());
-  }
-
-  List<Widget> _stateLegendSamples() {
-    return const [
-      _StateLegendSample(
-        kind: _TechLegendStateKind.researched,
-        state: _TechNodeState(
-          researched: true,
-          inProgress: false,
-          available: false,
-        ),
-      ),
-      _StateLegendSample(
-        kind: _TechLegendStateKind.inProgress,
-        state: _TechNodeState(
-          researched: false,
-          inProgress: true,
-          available: false,
-        ),
-      ),
-      _StateLegendSample(
-        kind: _TechLegendStateKind.available,
-        state: _TechNodeState(
-          researched: false,
-          inProgress: false,
-          available: true,
-        ),
-      ),
-      _StateLegendSample(
-        kind: _TechLegendStateKind.locked,
-        state: _TechNodeState(
-          researched: false,
-          inProgress: false,
-          available: false,
-        ),
-      ),
-    ];
-  }
-}
-
-class _LegendChip extends StatelessWidget {
-  const _LegendChip({required this.color, required this.label});
-
-  final Color color;
-  final String label;
-
-  @override
-  Widget build(BuildContext context) {
-    return Chip(
-      label: Text(label),
-      backgroundColor: color.withValues(alpha: 0.2),
-      side: BorderSide(color: color, width: 1.5),
-      labelStyle: Theme.of(context).textTheme.bodySmall,
-      visualDensity: VisualDensity.compact,
-      materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
-    );
-  }
-}
-
-class _StateLegendSample extends StatelessWidget {
-  const _StateLegendSample({required this.kind, required this.state});
-
-  final _TechLegendStateKind kind;
-  final _TechNodeState state;
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = appL10n(context);
-    // Use a dummy tech with neutral category just to render the style.
-    const dummyTech = TechDefinition(
-      id: 'legend_dummy',
-      era: 1,
-      category: 'gathering',
-      cost: 0,
-      prerequisiteIds: <String>[],
-      regimentUnlockIds: <String>[],
-      shipUnlockIds: <String>[],
-    );
-
-    return Row(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        SizedBox(
-          width: 72,
-          height: 24,
-          child: _TechNode(tech: dummyTech, state: state, onTap: () {}),
-        ),
-        const SizedBox(width: 4),
-        Text(
-          _localizedLabel(l10n),
-          style: Theme.of(context).textTheme.bodySmall,
-        ),
-      ],
-    );
-  }
-
-  String _localizedLabel(AppLocalizations l10n) {
-    return switch (kind) {
-      _TechLegendStateKind.researched => l10n.techTree_stateResearched,
-      _TechLegendStateKind.inProgress => l10n.techTree_stateInProgress,
-      _TechLegendStateKind.available => l10n.techTree_stateAvailable,
-      _TechLegendStateKind.locked => l10n.techTree_stateLocked,
-    };
   }
 }
